Convert all descriptions for content types to single `Wiki` objects
**NB** This branch is still broken. I have not edited the web interface
yet, just the back end. See next commit for completion

I had a problem with a proliferation of types: for each of the
following:

- Writer
- Text
- Edition
- Tag

I had a `db.Text` column called `bio`, `summary`, `history`, and
`description` (respectively). But when I eventually decide to introduce
edit history for all of those, I would have ended up with a
proliferation of types: `WriterEdit`, `TextEdit`, `EditionEdit`, etc. It
would have been ridiculous.

The solution was to create a simple type, `Wiki`, with a corresponding
type, `WikiEdit`. Each of those types no point to a `Wiki`. This means
I'll be able to create one route to edit any wiki.

I need to, perhaps, convert the Request system to use the same thing for
its various needs. But I'm not sure about that because they're much
different from these 4 types.

Regardless, this `Wiki` type is very scalable. I'll be able to use it
for essays and a lot of other things.

In the course of building this I made a few modifications to some other
stuff. It started with needing to abandon the sql files for `Text` and
`Writer`.

List of changes:

- `app/models.py` now has `Wiki` and `WikiEdit` and `Wiki` creation
  transparently creates a `WikiEdit`. Furthermore, creating any of the
  types that _use_ `Wiki` is just as transparent. Supply a `description`
  parameter to the construction of the object and a corresponding `Wiki`
  will be created, which will create a corresponding `WikiEdit`. I
  really need to build such transparency into the creation of an
  Annotation.
- `createdefaultusers.py` now has a `--communityonly` default. I want
  to convert this entire program to use a `yaml` file like
  `inserttags.py` now. I don't like having it all hardcoded anymore.
  Community will always be hardcoded. I will probably eliminate
  `--communityonly when I do this conversion and just default to
  creating community only.
- `insertlines.py` now requires a yaml file as an input (a template of
  which exists in the root structure of my `data/done` directory, which
  I would like to graft onto this git repo as a submodule eventually)
  that provides all the necessary information to create the `Text`,
  `Edition`, and writers, adding them as `authors` and connections. I'd
  been thinking for a while of doing exactly this sort of thing, but
  yaml turned out to be perfect for it.
- `insertannotations.py` just got debugged and cleaned up a bit is all.
- `inserttags.py` is now a new script that takes a yaml file and inserts
  new tags. This was necessary because of the new `Wiki` strategy.
- My bash scripts got adjusted to deal with new arguments and files.
  I also just cleaned up some of the output and adjusted them to require
  a password sooner.
- `scripts/sql/connection_enums.sql` got adjusted so that they can be
  used for the insertion of writer_connections in `insertlines.py`. It
  could have remained, but I didn't want it to. So instead I made them
  lower case so the yaml file could be to.
- Several sql files got removed since they're no longer necessary,
  including one that has been useless for a while.
- `scripts/tags.yml` is my default tags file now for `inserttags.py`.

+++ scripts/recreate.sh
@@ -1,14 +1,18 @@
 #!/bin/bash
+if [[ $# -eq 0 ]] ; then
+    echo "- need password for all default users."
+    exit 127
+fi
 set -e
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
-sudo echo "dropping and recreating database..."
+sudo echo "- dropping and recreating database..."
 echo "drop database icc; create database icc;" | sudo mysql icc
-echo "removing original migrations folder..."
+echo "- removing original migrations folder..."
 rm -rf $DIR/../migrations
-echo "creating new migrations folder..."
+echo "- creating new migrations folder..."
 /home/malan/projects/icc/icc/venv/bin/flask db init
-echo "recreating alembic scripts..."
+echo "- recreating alembic scripts..."
 /home/malan/projects/icc/icc/venv/bin/flask db migrate
-echo "recreating table structure..."
+echo "- recreating table structure..."
 /home/malan/projects/icc/icc/venv/bin/flask db upgrade
 $DIR/populate.sh $1
